ref(debuginfo): Separate breakpad sym parsing from symcache logic
